reorganizing gem, adding functionality to reference all files in gem from one main file
