+++ lib/message-contracts/base_message.rb
