+++ lib/message-contracts/topics/routing_keys.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-class Topic
+class RoutingKeys
   USER_CHANGED = "user.changed"
   USER_CREATED = "user.created"
   USER_SYNC = "user.sync"
