+++ lib/message-contracts/user_message.rb
