+++ lib/message-contracts/user_sync_message.rb
