+++ lib/message-contracts/version/version.rb
