--- lib/message/contracts.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-require_relative "contracts/version"
-
-module Message
-  module Contracts
-    class Error < StandardError; end
-    # Your code goes here...
-  end
-end
