+++ lib/message_contracts.rb
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+require_relative "message-contracts/base_message"
+require_relative "message-contracts/user_message"
+require_relative "message-contracts/user_sync_message"
+require_relative "message-contracts/topics/routing_keys"
+class MessageContracts
+end
