+++ message-contracts.gemspec
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require_relative "lib/message/contracts/version"
+require_relative "lib/message-contracts/version/version"
 
 Gem::Specification.new do |spec|
   spec.name          = "message-contracts"
