+++ spec/message-contracts/contracts_spec.rb
