+++ spec/spec_helper.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
 require "ostruct"
-require "message/contracts/base_message"
-require "message/contracts/user_message"
-require "message/contracts/user_sync_message"
+require "message-contracts/base_message"
+require "message-contracts/user_message"
+require "message-contracts/user_sync_message"
 
 RSpec.configure do |config|
   # Enable flags like --only-failures and --next-failure
